+++ src/fwup_apply.h
@@ -19,6 +19,7 @@
 
 #include "functions.h"
 
-int fwup_apply(const char *fw_filename, const char *task, const char *output_filename, enum fwup_apply_progress progress, const unsigned char *public_key);
+void fwup_apply_zero_progress(enum fwup_apply_progress progress);
+int fwup_apply(const char *fw_filename, const char *task, int output_fd, enum fwup_apply_progress progress, const unsigned char *public_key);
 
 #endif // FWUP_APPLY_H